Remove ParaView dependency for out-of-source builds
