+++ cmake/UseBioDynaMo.cmake.in
@@ -147,11 +147,7 @@ if(paraview)
   find_package(ParaView 5.3 QUIET OPTIONAL_COMPONENTS vtkPVPythonCatalyst)
   if(${ParaView_FOUND})
     message(STATUS "Found ParaView")
-    include("${PARAVIEW_USE_FILE}")
     add_definitions("-DUSE_CATALYST")
-    link_directories($ENV{ParaView_LIB_DIR})
-    set(BDM_REQUIRED_LIBRARIES ${BDM_REQUIRED_LIBRARIES} vtkPVPythonCatalyst
-        vtkIOParallelXML vtkPVVTKExtensionsDefault vtkPVServerManagerRendering)
   else()
     message(WARNING "ParaView not found")
     set(paraview OFF CACHE "Build with ParaView support." BOOL)
@@ -159,15 +155,13 @@ if(paraview)
 endif()
 
 # paraview version 5.5 (MacOS) does not contain TBB
-if (APPLE OR (NOT paraview OR NOT ${ParaView_FOUND}))
-  find_package(TBB REQUIRED)
-  if (TBB_FOUND)
-    include_directories(${TBB_INCLUDE_DIRS})
-    set(BDM_REQUIRED_LIBRARIES ${BDM_REQUIRED_LIBRARIES} ${TBB_LIBRARIES})
-  else()
-    MESSAGE(FATAL_ERROR "TBB was not found on your system."
-    " Please run ./prerequisites.sh again before calling cmake.")
-  endif()
+find_package(TBB REQUIRED)
+if (TBB_FOUND)
+  include_directories(${TBB_INCLUDE_DIRS})
+  set(BDM_REQUIRED_LIBRARIES ${BDM_REQUIRED_LIBRARIES} ${TBB_LIBRARIES})
+else()
+  MESSAGE(FATAL_ERROR "TBB was not found on your system."
+  " Please run ./prerequisites.sh again before calling cmake.")
 endif()
 
 find_package(ROOT REQUIRED COMPONENTS Geom Gui Eve GenVector)
