+++ src/core/visualization/catalyst_adaptor.cc
@@ -13,16 +13,361 @@
 // -----------------------------------------------------------------------------
 
 #include "core/visualization/catalyst_adaptor.h"
+#include "core/visualization/catalyst_helper.h"
+#include "core/visualization/insitu_pipeline.h"
 
 #if defined(USE_CATALYST) && !defined(__ROOTCLING__)
 
+#include <vtkCPDataDescription.h>
+#include <vtkCPInputDataDescription.h>
+#include <vtkCPProcessor.h>
+#include <vtkCPPythonScriptPipeline.h>
+#include <vtkDoubleArray.h>
+#include <vtkFieldData.h>
+#include <vtkIdTypeArray.h>
+#include <vtkImageData.h>
+#include <vtkIntArray.h>
+#include <vtkNew.h>
+#include <vtkPointData.h>
+#include <vtkPoints.h>
+#include <vtkStringArray.h>
+#include <vtkUnstructuredGrid.h>
+#include <vtkXMLPImageDataWriter.h>
+#include <vtkXMLPUnstructuredGridWriter.h>
+
 namespace bdm {
 
-vtkCPProcessor* CatalystAdaptor::g_processor_ = nullptr;
+struct CatalystAdaptor::ParaViewImpl {
+  vtkCPProcessor* g_processor_ = nullptr;
+  std::unordered_map<std::string, VtkSoGrid*> vtk_so_grids_;
+  std::unordered_map<std::string, VtkDiffusionGrid*> vtk_dgrids_;
+  InSituPipeline* pipeline_ = nullptr;
+  vtkCPDataDescription* data_description_ = nullptr;
+};
 
-constexpr const char* CatalystAdaptor::kSimulationInfoJson;
+CatalystAdaptor::CatalystAdaptor(const std::string& script)
+    : python_script_(script) {
+  counter_++;
+}
 
-std::atomic<uint64_t> CatalystAdaptor::counter_;
+CatalystAdaptor::~CatalystAdaptor() {
+  auto* param = Simulation::GetActive()->GetParam();
+  counter_--;
+
+  if (impl_) {
+    if (impl_->pipeline_) {
+      impl_->g_processor_->RemovePipeline(impl_->pipeline_);
+      impl_->pipeline_->Delete();
+      impl_->pipeline_ = nullptr;
+    }
+
+    if (counter_ == 0 && impl_->g_processor_) {
+      impl_->g_processor_->RemoveAllPipelines();
+      impl_->g_processor_->Finalize();
+      impl_->g_processor_->Delete();
+      impl_->g_processor_ = nullptr;
+    }
+    if (param->export_visualization_ &&
+        param->visualization_export_generate_pvsm_) {
+      GenerateSimulationInfoJson(impl_->vtk_so_grids_, impl_->vtk_dgrids_);
+      GenerateParaviewState();
+    }
+
+    for (auto& el : impl_->vtk_so_grids_) {
+      delete el.second;
+    }
+    for (auto& el : impl_->vtk_dgrids_) {
+      delete el.second;
+    }
+  }
+}
+
+void CatalystAdaptor::Visualize() {
+  if (!initialized_) {
+    Initialize();
+    initialized_ = true;
+  }
+
+  auto* sim = Simulation::GetActive();
+  auto* param = sim->GetParam();
+  uint64_t total_steps = sim->GetScheduler()->GetSimulatedSteps();
+  if (total_steps % param->visualization_export_interval_ != 0) {
+    return;
+  }
+
+  if (param->live_visualization_) {
+    double time = param->simulation_time_step_ * total_steps;
+    LiveVisualization(time, total_steps);
+  }
+  if (param->export_visualization_) {
+    double time = param->simulation_time_step_ * total_steps;
+    ExportVisualization(time, total_steps);
+  }
+}
+
+void CatalystAdaptor::Initialize() {
+  auto* sim = Simulation::GetActive();
+  auto* param = sim->GetParam();
+
+  exclusive_export_viz_ =
+      param->export_visualization_ && !param->live_visualization_;
+  if (param->live_visualization_ || param->export_visualization_) {
+    if (impl_->g_processor_ == nullptr) {
+      impl_->g_processor_ = vtkCPProcessor::New();
+      impl_->g_processor_->Initialize();
+    }
+
+    if (param->live_visualization_ &&
+        impl_->g_processor_->GetNumberOfPipelines() != 0) {
+      Log::Fatal("CatalystAdaptor",
+                 "Live visualization does not support multiple "
+                 "simulations. Turning off live visualization for ",
+                 sim->GetUniqueName());
+    } else if (param->python_catalyst_pipeline_) {
+      vtkNew<vtkCPPythonScriptPipeline> pipeline;
+      pipeline->Initialize(python_script_.c_str());
+      impl_->g_processor_->AddPipeline(pipeline.GetPointer());
+    } else if (!exclusive_export_viz_) {
+      impl_->pipeline_ = new InSituPipeline();
+      impl_->g_processor_->AddPipeline(impl_->pipeline_);
+    }
+
+    if (impl_->data_description_ == nullptr) {
+      impl_->data_description_ = vtkCPDataDescription::New();
+    } else {
+      impl_->data_description_->Delete();
+      impl_->data_description_ = vtkCPDataDescription::New();
+    }
+    impl_->data_description_->SetTimeData(0, 0);
+
+    auto* param = Simulation::GetActive()->GetParam();
+    for (auto& pair : param->visualize_sim_objects_) {
+      impl_->vtk_so_grids_[pair.first.c_str()] =
+          new VtkSoGrid(pair.first.c_str(), impl_->data_description_);
+    }
+    for (auto& entry : param->visualize_diffusion_) {
+      impl_->vtk_dgrids_[entry.name_] =
+          new VtkDiffusionGrid(entry.name_, impl_->data_description_);
+    }
+  }
+}
+
+void CatalystAdaptor::LiveVisualization(double time, size_t step) {
+  if (impl_->data_description_ == nullptr) {
+    impl_->data_description_ = vtkCPDataDescription::New();
+  } else {
+    impl_->data_description_->Delete();
+    impl_->data_description_ = vtkCPDataDescription::New();
+  }
+  impl_->data_description_->SetTimeData(time, step);
+
+  CreateVtkObjects();
+
+  if (impl_->pipeline_ != nullptr) {
+    if (!(impl_->pipeline_->IsInitialized())) {
+      impl_->pipeline_->Initialize(impl_->vtk_so_grids_);
+    }
+  }
+
+  impl_->g_processor_->CoProcess(impl_->data_description_);
+}
+
+/// Exports the visualized objects to file, so that they can be imported and
+/// visualized in ParaView at a later point in time
+///
+/// @param[in]  time            The simulation time
+/// @param[in]  step            The time step
+/// @param[in]  last_time_step  The last time step
+///
+void CatalystAdaptor::ExportVisualization(double time, size_t step) {
+  if (impl_->data_description_ == nullptr) {
+    impl_->data_description_ = vtkCPDataDescription::New();
+  } else {
+    impl_->data_description_->Delete();
+    impl_->data_description_ = vtkCPDataDescription::New();
+  }
+  impl_->data_description_->SetTimeData(time, step);
+
+  CreateVtkObjects();
+  WriteToFile(step);
+}
+
+void CatalystAdaptor::CreateVtkObjects() {
+  BuildSimObjectsVTKStructures();
+  BuildDiffusionGridVTKStructures();
+}
+
+void CatalystAdaptor::ProcessSimObject(const SimObject* so) {
+  auto* param = Simulation::GetActive()->GetParam();
+  auto so_name = so->GetTypeName();
+
+  if (param->visualize_sim_objects_.find(so_name) !=
+      param->visualize_sim_objects_.end()) {
+    // assert(impl_->vtk_so_grids_.find(so->GetTypeName()) !=
+    // impl_->vtk_so_grids_.end() &&
+    // Concat("VtkSoGrid for ", so->GetTypeName(), " not found!");
+    auto* vsg = impl_->vtk_so_grids_[so->GetTypeName()];
+    if (!vsg->initialized_) {
+      vsg->Init(so);
+    }
+
+    // If we segfault at here it probably means that there is a problem
+    // with the pipeline (either the C++ pipeline or Python pipeline)
+    // We do not need to RequestDataDescription in Export Mode, because
+    // we do not make use of Catalyst CoProcessing capabilities
+    if (exclusive_export_viz_ ||
+        (impl_->g_processor_->RequestDataDescription(
+            impl_->data_description_)) != 0) {
+      CatalystSoVisitor visitor(vsg);
+      so->ForEachDataMemberIn(vsg->vis_data_members_, &visitor);
+    }
+  }
+}
+
+void CatalystAdaptor::BuildSimObjectsVTKStructures() {
+  auto* rm = Simulation::GetActive()->GetResourceManager();
+
+  rm->ApplyOnAllElements([&](SimObject* so) { ProcessSimObject(so); });
+}
+
+// ---------------------------------------------------------------------------
+// diffusion grids
+
+void CatalystAdaptor::ProcessDiffusionGrid(const DiffusionGrid* grid) {
+  auto* param = Simulation::GetActive()->GetParam();
+  auto name = grid->GetSubstanceName();
+
+  // get visualization config
+  const Param::VisualizeDiffusion* vd = nullptr;
+  for (auto& entry : param->visualize_diffusion_) {
+    if (entry.name_ == name) {
+      vd = &entry;
+    }
+  }
+
+  if (vd != nullptr) {
+    auto* vdg = impl_->vtk_dgrids_[grid->GetSubstanceName()];
+    if (!vdg->used_) {
+      vdg->Init();
+    }
+
+    // If we segfault at here it probably means that there is a problem
+    // with  the pipeline (either the C++ pipeline or Python pipeline)
+    // We do not need to RequestDataDescription in Export Mode, because
+    // we do not make use of Catalyst CoProcessing capabilities
+    if (exclusive_export_viz_ ||
+        (impl_->g_processor_->RequestDataDescription(
+            impl_->data_description_)) != 0) {
+      auto num_boxes = grid->GetNumBoxesArray();
+      auto grid_dimensions = grid->GetDimensions();
+      auto box_length = grid->GetBoxLength();
+      auto total_boxes = grid->GetNumBoxes();
+
+      double origin_x = grid_dimensions[0];
+      double origin_y = grid_dimensions[2];
+      double origin_z = grid_dimensions[4];
+      vdg->data_->SetOrigin(origin_x, origin_y, origin_z);
+      vdg->data_->SetDimensions(num_boxes[0], num_boxes[1], num_boxes[2]);
+      vdg->data_->SetSpacing(box_length, box_length, box_length);
+
+      if (vdg->concentration_) {
+        auto* co_ptr = const_cast<double*>(grid->GetAllConcentrations());
+        vdg->concentration_->SetArray(co_ptr,
+                                      static_cast<vtkIdType>(total_boxes), 1);
+      }
+      if (vdg->gradient_) {
+        auto gr_ptr = const_cast<double*>(grid->GetAllGradients());
+        vdg->gradient_->SetArray(gr_ptr,
+                                 static_cast<vtkIdType>(total_boxes * 3), 1);
+      }
+    }
+  }
+}
+
+void CatalystAdaptor::BuildDiffusionGridVTKStructures() {
+  auto* rm = Simulation::GetActive()->GetResourceManager();
+
+  rm->ApplyOnAllDiffusionGrids(
+      [&](DiffusionGrid* grid) { ProcessDiffusionGrid(grid); });
+}
+
+void CatalystAdaptor::WriteToFile(size_t step) {
+  auto* sim = Simulation::GetActive();
+  for (auto& el : impl_->vtk_so_grids_) {
+    vtkNew<vtkXMLPUnstructuredGridWriter> cells_writer;
+    auto filename =
+        Concat(sim->GetOutputDir(), "/", el.second->name_, "-", step, ".pvtu");
+    cells_writer->SetFileName(filename.c_str());
+    cells_writer->SetInputData(el.second->data_);
+    cells_writer->Update();
+  }
+
+  for (auto& entry : impl_->vtk_dgrids_) {
+    vtkNew<vtkXMLPImageDataWriter> dgrid_writer;
+
+    const auto& substance_name = entry.second->name_;
+    auto filename =
+        Concat(sim->GetOutputDir(), "/", substance_name, "-", step, ".pvti");
+    dgrid_writer->SetFileName(filename.c_str());
+    dgrid_writer->SetInputData(entry.second->data_);
+    dgrid_writer->Update();
+  }
+}
+
+/// This function generates the Paraview state based on the exported files
+/// Therefore, the user can load the visualization simply by opening the pvsm
+/// file and does not have to perform a lot of manual steps.
+void CatalystAdaptor::GenerateParaviewState() {
+  auto* sim = Simulation::GetActive();
+  std::stringstream python_cmd;
+  std::string bdm_src_dir = std::getenv("BDM_SRC_DIR");
+
+  python_cmd << bdm_src_dir << "/../third_party/paraview/bin/pvpython "
+             << bdm_src_dir << "/core/visualization/generate_pv_state.py "
+             << sim->GetOutputDir() << "/" << kSimulationInfoJson;
+  int ret_code = system(python_cmd.str().c_str());
+  if (ret_code) {
+    Log::Fatal("CatalystAdaptor::GenerateParaviewState",
+               "Error during generation of ParaView state\n", "Command\n",
+               python_cmd.str());
+  }
+}
+
+}  // namespace bdm
+
+#else
+
+#include <string>
+#include <unordered_map>
+#include "core/shape.h"
+
+namespace bdm {
+
+/// False front (to ignore Catalyst in gtests)
+class CatalystAdaptor {
+ public:
+  explicit CatalystAdaptor(const std::string& script) {}
+
+  void Visualize() {}
+
+ private:
+  friend class CatalystAdaptorTest_GenerateSimulationInfoJson_Test;
+  friend class CatalystAdaptorTest_GenerateParaviewState_Test;
+  friend class CatalystAdaptorTest_CheckVisualizationSelection_Test;
+  friend class DISABLED_DiffusionTest_ModelInitializer_Test;
+
+  void LiveVisualization(double time, size_t time_step) {}
+
+  void ExportVisualization(double step, size_t time_step) {}
+
+  void WriteToFile(size_t step) {}
+
+  static void GenerateParaviewState() {}
+};
+
+void GenerateSimulationInfoJson(
+    const std::unordered_map<std::string, VtkSoGrid*>& vtk_so_grids,
+    const std::unordered_map<std::string, VtkDiffusionGrid*>& vtk_dgrids) {}
 
 }  // namespace bdm
